requirements.txt
PySide6>=6.0.0
Flask>=2.0.0
Pillow>=9.0.0 